Line Count Programming Challenge
Here's a simple help free challenge to get you started: write a program that takes a file as an argument and counts the total number of lines. Lines are defined as ending with a newline character. Program usage should be

count.exe filename.txt

and the output should be the line count.